requirements.txt
asgiref==3.4.1
Django==4.0.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
mock==4.0.3
numpy==1.22.1
openpyxl==3.0.9
pandas==1.3.5
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
